<wh> integrasi dengan raja ongkir
perubahan pada konsep pengambilan provinsi, kota dan kecamatan disesuaikan dengan dari rajaongkir
penambahan fasilitas traking baik di core maupun di schedule (setiap 6 jam)
penambahan api untuk pelacakan produk

+++ database/migrations/2021_07_07_193656_create_bimbingan_materi_table.php
@@ -14,7 +14,7 @@ class CreateBimbinganMateriTable extends Migration
     public function up()
     {
         Schema::create('bimbingan_materi', function (Blueprint $table) {
-            $table->integer('bimbingan_id')->unsigned();
+            $table->bigInteger('bimbingan_id')->unsigned();
             $table->foreign('bimbingan_id')->references('id')->on('bimbingan');
             $table->integer('materi_id')->unsigned();
             $table->foreign('materi_id')->references('id')->on('materi');
